lab5/requirements.txt
pycryptodome==3.23.0
pycryptodomex==3.23.0
